rename: change directory of model list api response dto
- move ModelListResponseDto, Model to api/modelList/dto/response directory
