+++ src/main/java/kr/co/ipalab/lingobe/api/modelList/dto/response/Model.java
@@ -1,6 +1,5 @@
-package kr.co.ipalab.lingobe.flask.modelList.dto;
+package kr.co.ipalab.lingobe.api.modelList.dto.response;
 
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,6 +9,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ModelListResponseDto {
-    private List<Model> modelList;
+public class Model {
+    private String name;
 }
