+++ src/main/java/kr/co/ipalab/lingobe/api/modelList/dto/response/ModelListResponseDto.java
@@ -1,5 +1,6 @@
-package kr.co.ipalab.lingobe.flask.modelList.dto;
+package kr.co.ipalab.lingobe.api.modelList.dto.response;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,6 +10,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Model {
-    private String name;
+public class ModelListResponseDto {
+    private List<Model> modelList;
 }
